+++ src/generation/sass-graph/index.ts
@@ -4,10 +4,15 @@ import { Graph } from "../../graph";
 import { Path } from "../../path";
 import { sassGraphGraphToGraph } from "./translator";
 
-export function generateGraphFromSassGraph(rootFolder: Path): Graph {
+export function generateGraphFromSassGraph(
+  rootFolder: Path,
+  loadPaths: Path[],
+): Graph {
   return sassGraphGraphToGraph(
     sassGraph.parseDir(rootFolder, {
       extensions: ["sass", "scss", "less", "css"],
+      // @ts-ignore There's a typo in sass-graph types
+      loadPaths,
     }),
   );
 }